Novel_Rewrite_App/requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.0.1
pytesseract==0.3.10
Werkzeug==2.3.7 